fix(discovery): keep rotation ages strongly typed

+++ lib/data/services/seerr/seerr_discovery_rotation_history.dart
@@ -37,7 +37,8 @@ class SeerrDiscoveryRotationHistory {
   /// ago, and so on.
   Map<String, int> get sessionsSinceSeen => {
         for (final entry in _lastSeenSession.entries)
-          entry.key: (_sessionNumber - 1 - entry.value).clamp(0, 1 << 30),
+          entry.key:
+              (_sessionNumber - 1 - entry.value).clamp(0, 1 << 30).toInt(),
       };
 
   void commitSession(Iterable<String> selectedSectionIds) {
